refactor(kinnex-report): moves kinex report structure and fetch functions to lib dir
