+++ tests/unit/views/SampleReport.spec.js
@@ -1,8 +1,21 @@
 import { mount } from '@support/testHelper'
 import useRootStore from '@/stores'
 import SampleReport from '@/views/SampleReport.vue'
+import { createPinia, setActivePinia, flushPromises } from '@support/testHelper.js'
+import { fetchTractionSamples, fetchSequencescapeSamples } from '@/lib/reports/KinnexReport.js'
 
 const mockShowAlert = vi.fn()
+
+// We need to hoist and mock the fetch functions to control their behavior in tests
+// and also keep the original functionality of the rest of the module
+// See https://vitest.dev/api/vi.html#vi-mock
+const mocks = vi.hoisted(() => {
+  return {
+    fetchSequencescapeSamples: vi.fn(),
+    fetchTractionSamples: vi.fn(),
+  }
+})
+
 vi.mock('@/composables/useAlert', () => ({
   default: () => ({
     showAlert: mockShowAlert,
@@ -15,149 +28,42 @@ vi.mock('@/lib/csv/creator', async (importOriginal) => {
     downloadBlob: vi.fn(), // mock the downloadBlob method
   }
 })
-import { createPinia, setActivePinia, flushPromises } from '@support/testHelper.js'
+vi.mock('@/lib/reports/KinnexReport.js', async (importOriginal) => {
+  const actual = await importOriginal()
+  return {
+    ...actual,
+    fetchSequencescapeSamples: mocks.fetchSequencescapeSamples,
+    fetchTractionSamples: mocks.fetchTractionSamples,
+  }
+})
+
+const tractionExpectedSample = {
+  request_id: 1,
+  cost_code: 'aCostCodeExample',
+  library_type: 'Pacbio_HiFi',
+  date_of_sample_collection: '2025-06-23',
+  supplier_name: 'Supplier B',
+  donor_id: 'donor-123',
+  species: 'human',
+  external_id: '123',
+}
+
+const ssExpectedSample = {
+  id: '2',
+  sanger_sample_id: 'id-123',
+  uuid: '123',
+  cohort: 'Cohort 1',
+  concentration: 50,
+  volume: 100,
+  study_number: '1',
+  study_name: 'Study 1',
+  submitting_faculty: 'Faculty Sponsor 1',
+}
 
 describe('SampleReport', () => {
   let rootStore, get, ss_get
   const buildWrapper = () => {
     return mount(SampleReport)
-  }
-
-  // TODO: hypothetical sample data for testing - move to factories or fixtures
-  const request = {
-    id: 1,
-    type: 'requests',
-    attributes: {
-      library_type: 'Pacbio_HiFi',
-      estimate_of_gb_required: null,
-      number_of_smrt_cells: null,
-      cost_code: 'aCostCodeExample',
-      external_study_id: '123',
-      sample_name: 'Supplier B',
-      barcode: 'NT6746',
-      sample_species: 'human',
-      created_at: '2025/06/23 13:05',
-      source_identifier: 'NT6746',
-      sample_retention_instruction: null,
-    },
-    relationships: {
-      sample: {
-        data: {
-          type: 'samples',
-          id: '2',
-        },
-      },
-    },
-  }
-
-  const sample = {
-    id: '2',
-    type: 'samples',
-    attributes: {
-      name: 'test-sample-1',
-      date_of_sample_collection: '2025-06-23',
-      donor_id: 'donor-123',
-      species: 'human',
-      external_id: '123',
-      retention_instruction: null,
-      supplier_name: 'Supplier B',
-      sanger_sample_id: 'id-123',
-    },
-  }
-
-  const ss_sample = {
-    id: '2',
-    type: 'samples',
-    attributes: {
-      uuid: '123',
-      sanger_sample_id: 'id-123',
-    },
-    relationships: {
-      sample_metadata: {
-        data: {
-          type: 'sample_metadata',
-          id: '1',
-        },
-      },
-      studies: {
-        data: [
-          {
-            type: 'studies',
-            id: '1',
-          },
-        ],
-      },
-    },
-  }
-
-  const ss_study = {
-    id: '1',
-    type: 'studies',
-    attributes: {
-      name: 'Study 1',
-    },
-    relationships: {
-      study_metadata: {
-        data: {
-          type: 'study_metadata',
-          id: '1',
-        },
-      },
-    },
-  }
-
-  const ss_studyMetadata = {
-    id: '1',
-    type: 'study_metadata',
-    relationships: {
-      faculty_sponsor: {
-        data: {
-          type: 'faculty_sponsors',
-          id: '1',
-        },
-      },
-    },
-  }
-
-  const ss_sampleMetadata = {
-    id: '1',
-    type: 'sample_metadata',
-    attributes: {
-      cohort: 'Cohort 1',
-      concentration: 50,
-      volume: 100,
-    },
-  }
-
-  const ss_facultySponsor = {
-    id: '1',
-    type: 'faculty_sponsors',
-    attributes: {
-      name: 'Faculty Sponsor 1',
-    },
-  }
-
-  const expectedSample = {
-    // SS Attributes
-    id: ss_sample.id,
-    sanger_sample_id: ss_sample.attributes.sanger_sample_id,
-    uuid: ss_sample.attributes.uuid,
-    cohort: ss_sampleMetadata.attributes.cohort,
-    concentration: ss_sampleMetadata.attributes.concentration,
-    volume: ss_sampleMetadata.attributes.volume,
-    study_number: ss_study.id,
-    study_name: ss_study.attributes.name,
-    submitting_faculty: ss_facultySponsor.attributes.name,
-
-    // Traction attributes
-    request_id: request.id,
-    cost_code: request.attributes.cost_code,
-    library_type: request.attributes.library_type,
-    date_of_sample_collection: sample.attributes.date_of_sample_collection,
-    supplier_name: sample.attributes.supplier_name,
-    donor_id: sample.attributes.donor_id,
-    species: sample.attributes.species,
-    external_id: sample.attributes.external_id,
   }
 
   beforeEach(() => {
@@ -178,121 +84,84 @@ describe('SampleReport', () => {
   describe('sample input', () => {
     it('adds a sample when the input is not empty', async () => {
       const wrapper = buildWrapper()
-      get.mockResolvedValueOnce({
-        status: 200,
-        statusText: 'OK',
-        ok: true,
-        json: () => Promise.resolve({ data: [request], included: [sample] }),
-      })
-      // Second call is for source_identifier fetch
-      get.mockResolvedValueOnce({
-        status: 200,
-        statusText: 'OK',
-        ok: true,
-        json: () => Promise.resolve({ data: [], included: [] }),
-      })
-      ss_get.mockResolvedValue({
-        status: 200,
-        statusText: 'OK',
-        ok: true,
-        json: () =>
-          Promise.resolve({
-            data: [ss_sample],
-            included: [ss_facultySponsor, ss_sampleMetadata, ss_studyMetadata, ss_study],
-          }),
-      })
+      vi.mocked(fetchTractionSamples).mockResolvedValue([tractionExpectedSample])
+      vi.mocked(fetchSequencescapeSamples).mockResolvedValue([ssExpectedSample])
 
       const sampleInput = wrapper.find('#sampleInput')
       await sampleInput.setValue('sample1')
       await wrapper.find('#searchSamples').trigger('click')
       await flushPromises()
 
-      expect(wrapper.vm.samples).toEqual([expectedSample])
+      expect(wrapper.vm.samples).toEqual([{ ...tractionExpectedSample, ...ssExpectedSample }])
     })
 
-    it('does not add a sample when the service returns no data', async () => {
+    it('correctly combines multiple samples', async () => {
       const wrapper = buildWrapper()
-      get.mockResolvedValue({
-        status: 200,
-        statusText: 'OK',
-        ok: true,
-        json: () => Promise.resolve({ data: [], included: [] }),
-      })
+      vi.mocked(fetchTractionSamples).mockResolvedValue([
+        tractionExpectedSample,
+        // only include information required for linking and some metadata
+        { cost_code: 'cost-code-2', species: 'species-1', external_id: 'sample2-id' },
+        { cost_code: 'cost-code-3', species: 'species-2', external_id: 'sample3-id' },
+        // Sample with no matching Sequencescape data
+        { cost_code: 'cost-code-4', species: 'species-3', external_id: 'sample4-id' },
+      ])
+      vi.mocked(fetchSequencescapeSamples).mockResolvedValue([
+        ssExpectedSample,
+        // only include information required for linking and some metadata
+        { concentration: 5, volume: 10, uuid: 'sample2-id' },
+        { concentration: 1, volume: 1, uuid: 'sample3-id' },
+      ])
 
       const sampleInput = wrapper.find('#sampleInput')
-      await sampleInput.setValue('123')
+      await sampleInput.setValue('sample1,sample2,sample3')
+      await wrapper.find('#searchSamples').trigger('click')
+      await flushPromises()
+
+      expect(wrapper.vm.samples).toEqual([
+        { ...tractionExpectedSample, ...ssExpectedSample },
+        {
+          cost_code: 'cost-code-2',
+          species: 'species-1',
+          external_id: 'sample2-id',
+          concentration: 5,
+          volume: 10,
+          uuid: 'sample2-id',
+        },
+        {
+          cost_code: 'cost-code-3',
+          species: 'species-2',
+          external_id: 'sample3-id',
+          concentration: 1,
+          volume: 1,
+          uuid: 'sample3-id',
+        },
+        { cost_code: 'cost-code-4', species: 'species-3', external_id: 'sample4-id' },
+      ])
+    })
+
+    it('does not add a sample if traction returns empty', async () => {
+      const wrapper = buildWrapper()
+      vi.mocked(fetchTractionSamples).mockResolvedValue([])
+
+      const sampleInput = wrapper.find('#sampleInput')
+      await sampleInput.setValue('sample1')
       await wrapper.find('#searchSamples').trigger('click')
       await flushPromises()
 
       expect(wrapper.vm.samples).toEqual([])
-      expect(mockShowAlert).toBeCalledWith(
-        'No samples found in Traction with the provided input',
-        'warning',
-      )
     })
 
-    it('does not add a sample if it already exists in the list', async () => {
+    it('does not add a sample if traction returns but sequencescape does not', async () => {
       const wrapper = buildWrapper()
-      get.mockResolvedValueOnce({
-        status: 200,
-        statusText: 'OK',
-        ok: true,
-        json: () => Promise.resolve({ data: [request], included: [sample] }),
-      })
-      // Second call is for source_identifier fetch
-      get.mockResolvedValueOnce({
-        status: 200,
-        statusText: 'OK',
-        ok: true,
-        json: () => Promise.resolve({ data: [], included: [] }),
-      })
-      wrapper.vm.samples = [{ request_id: request.id }]
+      vi.mocked(fetchTractionSamples).mockResolvedValue([tractionExpectedSample])
+      vi.mocked(fetchSequencescapeSamples).mockResolvedValue([])
 
       const sampleInput = wrapper.find('#sampleInput')
       await sampleInput.setValue('sample1')
       await wrapper.find('#searchSamples').trigger('click')
       await flushPromises()
 
-      expect(wrapper.vm.samples).toEqual([{ request_id: request.id }])
-      expect(mockShowAlert).toBeCalledWith('Sample sample1 already exists in the list', 'info')
-    })
-
-    it('shows an error alert if the service returns an error', async () => {
-      const wrapper = buildWrapper()
-      get.mockRejectedValue('Internal Server Error')
-
-      const sampleInput = wrapper.find('#sampleInput')
-      await sampleInput.setValue('errorSample')
-      await wrapper.find('#searchSamples').trigger('click')
-      await flushPromises()
-
-      expect(mockShowAlert).toBeCalledWith(
-        'Error fetching samples from Traction: Internal Server Error',
-        'danger',
-      )
-    })
-
-    it('shows an error alert if sequencescape returns an error', async () => {
-      const wrapper = buildWrapper()
-      // Traction success
-      get.mockResolvedValue({
-        status: 200,
-        statusText: 'OK',
-        ok: true,
-        json: () => Promise.resolve({ data: [request], included: [sample] }),
-      })
-      // Sequencescape error
-      ss_get.mockRejectedValue('Internal Server Error')
-
-      const sampleInput = wrapper.find('#sampleInput')
-      await sampleInput.setValue('errorSample')
-      await wrapper.find('#searchSamples').trigger('click')
-      await flushPromises()
-
-      expect(mockShowAlert).toBeCalledWith(
-        'Error fetching samples from Sequencescape: Internal Server Error',
-        'danger',
-      )
+      expect(wrapper.vm.samples).toEqual([])
     })
   })
 
@@ -315,7 +184,7 @@ describe('SampleReport', () => {
 
     it('shows the correct sample data', async () => {
       const wrapper = buildWrapper()
-      wrapper.vm.samples = [{ id: request.id }, { id: 2 }]
+      wrapper.vm.samples = [{ id: 1 }, { id: 2 }]
       await wrapper.vm.$nextTick()
       const rows = wrapper.findAll('tr')
       expect(rows.length).toBe(3) // 2 samples + header row
@@ -323,7 +192,7 @@ describe('SampleReport', () => {
 
     it('removes the correct row on remove button click', async () => {
       const wrapper = buildWrapper()
-      wrapper.vm.samples = [{ id: request.id }, { id: 2 }]
+      wrapper.vm.samples = [{ id: 1 }, { id: 2 }]
       await wrapper.vm.$nextTick()
       const rows = wrapper.findAll('tr')
       expect(wrapper.vm.samples.length).toBe(2)
@@ -340,7 +209,7 @@ describe('SampleReport', () => {
     describe('Reset', () => {
       it('resets the sample list when reset button is clicked', async () => {
         const wrapper = buildWrapper()
-        wrapper.vm.samples = [{ request_id: request.id }, { request_id: 2 }]
+        wrapper.vm.samples = [{ request_id: 1 }, { request_id: 2 }]
         await wrapper.vm.$nextTick()
         expect(wrapper.vm.samples.length).toBe(2)
 
@@ -359,7 +228,7 @@ describe('SampleReport', () => {
 
       it('is enabled when samples are present', async () => {
         const wrapper = buildWrapper()
-        wrapper.vm.samples = [{ id: request.id }]
+        wrapper.vm.samples = [{ id: 1 }]
         await wrapper.vm.$nextTick()
         expect(wrapper.find('#download').element.disabled).toBe(false)
       })
@@ -370,7 +239,7 @@ describe('SampleReport', () => {
     describe('downloadReport', () => {
       it('downloads a CSV file with the correct content', async () => {
         const wrapper = buildWrapper()
-        wrapper.vm.samples = [expectedSample]
+        wrapper.vm.samples = [{ ...tractionExpectedSample, ...ssExpectedSample }]
         await wrapper.vm.$nextTick()
 
         // Get the mocked downloadBlob from the module mock
@@ -394,7 +263,7 @@ describe('SampleReport', () => {
     describe('reset', () => {
       it('clears the sample list', () => {
         const wrapper = buildWrapper()
-        wrapper.vm.samples = [expectedSample]
+        wrapper.vm.samples = [{ ...tractionExpectedSample, ...ssExpectedSample }]
         expect(wrapper.vm.samples.length).toBe(1)
         wrapper.vm.reset()
         expect(wrapper.vm.samples.length).toBe(0)
@@ -404,197 +273,18 @@ describe('SampleReport', () => {
     describe('removeSample', () => {
       it('removes the sample given the request_id', () => {
         const wrapper = buildWrapper()
-        wrapper.vm.samples = [{ request_id: request.id }, { request_id: 2 }]
+        wrapper.vm.samples = [{ request_id: 1 }, { request_id: 2 }]
         expect(wrapper.vm.samples.length).toBe(2)
 
-        wrapper.vm.removeSample(request.id)
+        wrapper.vm.removeSample(1)
         expect(wrapper.vm.samples.length).toBe(1)
         expect(wrapper.vm.samples[0].request_id).toBe(2)
       })
 
       it('does not throw an error if the index is out of bounds', () => {
         const wrapper = buildWrapper()
-        wrapper.vm.samples = [{ request_id: request.id }]
+        wrapper.vm.samples = [{ request_id: 1 }]
         expect(() => wrapper.vm.removeSample(5)).not.toThrow()
-      })
-    })
-
-    describe('fetchTractionSamples', () => {
-      it('fetches samples from Traction API', async () => {
-        const wrapper = buildWrapper()
-        wrapper.vm.sample_input = 'sample1'
-        get.mockResolvedValueOnce({
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-          json: () => Promise.resolve({ data: [request], included: [sample] }),
-        })
-        // Second call is for source_identifier fetch
-        get.mockResolvedValueOnce({
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-          json: () => Promise.resolve({ data: [], included: [] }),
-        })
-
-        const tractionSamples = await wrapper.vm.fetchTractionSamples()
-        expect(tractionSamples).toEqual([
-          {
-            request_id: request.id,
-            cost_code: request.attributes.cost_code,
-            date_of_sample_collection: sample.attributes.date_of_sample_collection,
-            donor_id: sample.attributes.donor_id,
-            library_type: request.attributes.library_type,
-            species: sample.attributes.species,
-            supplier_name: sample.attributes.supplier_name,
-            external_id: sample.attributes.external_id,
-          },
-        ])
-        expect(get).toHaveBeenCalledWith({
-          filter: { sample_name: 'sample1' },
-          include: 'sample',
-        })
-      })
-
-      it('fetches multiple samples from Traction API across filter types', async () => {
-        const wrapper = buildWrapper()
-        const sample2 = {
-          ...sample,
-          id: '3',
-          attributes: { ...sample.attributes, name: 'sample2', donor_id: 'another-id' },
-        }
-        const request2 = {
-          ...request,
-          id: 2,
-          attributes: {
-            ...request.attributes,
-            sample_name: 'sample2',
-            supplier_name: 'Supplier C',
-            library_type: 'PacBio-2',
-          },
-          relationships: {
-            sample: {
-              data: {
-                type: 'samples',
-                id: '3',
-              },
-            },
-          },
-        }
-        wrapper.vm.sample_input = 'sample1,sample2'
-        get.mockResolvedValueOnce({
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-          json: () => Promise.resolve({ data: [request], included: [sample] }),
-        })
-        // Second call is for source_identifier fetch
-        get.mockResolvedValueOnce({
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-          json: () => Promise.resolve({ data: [request2], included: [sample2] }),
-        })
-        const tractionSamples = await wrapper.vm.fetchTractionSamples()
-        expect(tractionSamples).toEqual([
-          {
-            request_id: request.id,
-            cost_code: request.attributes.cost_code,
-            date_of_sample_collection: sample.attributes.date_of_sample_collection,
-            donor_id: sample.attributes.donor_id,
-            library_type: request.attributes.library_type,
-            species: sample.attributes.species,
-            supplier_name: sample.attributes.supplier_name,
-            external_id: sample.attributes.external_id,
-          },
-          {
-            request_id: request2.id,
-            cost_code: request2.attributes.cost_code,
-            date_of_sample_collection: sample2.attributes.date_of_sample_collection,
-            donor_id: sample2.attributes.donor_id,
-            library_type: request2.attributes.library_type,
-            species: sample2.attributes.species,
-            supplier_name: sample2.attributes.supplier_name,
-            external_id: sample2.attributes.external_id,
-          },
-        ])
-        expect(get).toHaveBeenCalledWith({
-          filter: { sample_name: 'sample1,sample2' },
-          include: 'sample',
-        })
-      })
-
-      it('shows an error alert if Traction API returns an error', async () => {
-        const wrapper = buildWrapper()
-        get.mockRejectedValue('Internal Server Error')
-
-        await wrapper.vm.fetchTractionSamples()
-        expect(mockShowAlert).toBeCalledWith(
-          'Error fetching samples from Traction: Internal Server Error',
-          'danger',
-        )
-      })
-
-      it('shows a warning if the sample already exists', async () => {
-        const wrapper = buildWrapper()
-        wrapper.vm.samples = [{ request_id: request.id }]
-        wrapper.vm.sample_input = 'sample1'
-        get.mockResolvedValue({
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-          json: () => Promise.resolve({ data: [request], included: [sample] }),
-        })
-
-        await wrapper.vm.fetchTractionSamples()
-        expect(mockShowAlert).toBeCalledWith('Sample sample1 already exists in the list', 'info')
-      })
-    })
-
-    describe('fetchSequencescapeSamples', () => {
-      it('fetches samples from Sequencescape API', async () => {
-        const wrapper = buildWrapper()
-        wrapper.vm.samples = [expectedSample]
-        ss_get.mockResolvedValue({
-          status: 200,
-          statusText: 'OK',
-          ok: true,
-          json: () =>
-            Promise.resolve({
-              data: [ss_sample],
-              included: [ss_facultySponsor, ss_sampleMetadata, ss_studyMetadata, ss_study],
-            }),
-        })
-
-        const sequencescapeSamples = await wrapper.vm.fetchSequencescapeSamples(wrapper.vm.samples)
-        expect(sequencescapeSamples).toEqual([
-          {
-            id: ss_sample.id,
-            sanger_sample_id: ss_sample.attributes.sanger_sample_id,
-            uuid: ss_sample.attributes.uuid,
-            cohort: ss_sampleMetadata.attributes.cohort,
-            concentration: ss_sampleMetadata.attributes.concentration,
-            volume: ss_sampleMetadata.attributes.volume,
-            study_number: ss_study.id,
-            study_name: ss_study.attributes.name,
-            submitting_faculty: ss_facultySponsor.attributes.name,
-          },
-        ])
-        expect(ss_get).toHaveBeenCalledWith({
-          filter: { uuid: expectedSample.external_id },
-          include: 'sample_metadata,studies.study_metadata.faculty_sponsor',
-        })
-      })
-
-      it('shows an error alert if Sequencescape API returns an error', async () => {
-        const wrapper = buildWrapper()
-        ss_get.mockRejectedValue('Internal Server Error')
-
-        await wrapper.vm.fetchSequencescapeSamples([expectedSample])
-        expect(mockShowAlert).toBeCalledWith(
-          'Error fetching samples from Sequencescape: Internal Server Error',
-          'danger',
-        )
       })
     })
   })
